model_inference/fastspeech2_melgan/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
audioread==3.0.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
clang==5.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dataclasses==0.6
decorator==5.1.1
Distance==0.1.3
docopt==0.6.2
filelock==3.10.0
fire==0.5.0
flatbuffers==1.12
fonttools==4.39.2
g2p-en==2.1.0
g2pM==0.1.2.5
gast==0.4.0
german-transliterate @ git+https://github.com/repodiac/german_transliterate.git@4e01beba5f19adb9ccd72d0607a81f4ed7f6ee87
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.1.0
huggingface-hub==0.13.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
inflect==6.0.2
jamo==0.4.1
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.31.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mpmath==1.3.0
nltk==3.8.1
num2words==0.5.12
numba==0.48.0
numpy==1.23.0
oauthlib==3.2.2
onnx==1.13.1
onnxconverter-common==1.13.0
onnxruntime==1.14.1
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.7.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
pyparsing==3.0.9
pypinyin==0.48.0
python-dateutil==2.8.2
pyworld==0.3.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.15.0
soundfile==0.12.1
sympy==1.11.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
TensorFlowTTS==1.8
termcolor==1.1.0
TextGrid==1.5
tf-estimator-nightly==2.8.0.dev2021122109
tf2onnx==1.13.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.27.1
typeguard==2.13.3
typing-extensions==3.7.4.3
Unidecode==1.3.6
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.12.1
zipp==3.15.0
